Fixed Physics, Rebound Jump, and Rebuilt Test Area
Found and fixed the physics bug that caused the player to get snapped
off the edges of platforms. Since implementing this fix, I've noticed
what appears to be some minor clipping sometimes taking place when
hitting the bottom corner of a block while falling diagonally, but
I haven't been able to reproduce it to attempt a fix.
I also rebound the jump button to fire1, and rebuilt the testing area so
as to make it more useful for testing collision.