Semisolids, Scrolling, and 1 Block Gap Physics
Introduced semisolid platforms to the game, introduced screen scrolling,
added special code in the physics to allow the player to fit into
vertical 1 block gaps while jumping, and reworked the test map once
again.

Known issue: Jumping into the bottom-left corner of a block appears to
reliably clip the player inside of it. I believe the issue requires
a pixel perfect approach to the block.