Collision Bugfixes
Fixed two major issues with the collision detection code that resulted
in the mishandling of almost every collision check. By mere coincidence,
the map was designed in such a way that these issues never changed the
result of a collision check, but their effects would have been
immediately noticeable with even minor changes.